requirements.txt
numpy==1.14.5
appnope==0.1.0
bleach==1.5.0
click==6.7
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12.2
Flask-API==1.0
flask-restframework==0.2.1
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.2
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.4.1
olefile==0.44
opencv-python
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
Pillow-PIL==0.1.dev0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
waitress==1.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
